RFC200: Move compiler compliance from 1.5 to 1.4.

+++ org.osgi.impl.service.resourcemanagement/src/org/osgi/impl/service/resourcemanagement/ResourceContextEventNotifierImpl.java
@@ -117,7 +117,7 @@ public class ResourceContextEventNotifierImpl implements
 			filter = (int[]) propertyValue;
 		} catch (ClassCastException e) {
 			// should be an int
-			filter = new int[] {(Integer) propertyValue};
+			filter = new int[] {((Integer) propertyValue).intValue()};
 		}
 
 		return filter;
